backend/requirements/dev.txt
-r base.txt

pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
ipython==8.18.1
django-debug-toolbar==4.2.0
